requirements.txt
django-phonenumber-field
django-jazzmin
Django
djangorestframework
django-cors-headers
